improve scheduling a match & make bot cleanup properly after the admin deleted a channel

+++ queries/streamers.sql
@@ -1,23 +1,23 @@
 -- name: AddMatchStreamer :exec
 INSERT INTO streamers (
     channel_id,
-    name,
+    user_id,
     url
 ) VALUES (
     :channel_id,
-    :name,
+    :user_id,
     :url
 );
 
 -- name: DeleteMatchStreamer :exec
 DELETE FROM streamers
 WHERE channel_id = :channel_id
-AND name = :name;
+AND user_id = :user_id;
 
 -- name: DeleteMatchStreamers :exec
 DELETE FROM streamers
 WHERE channel_id = :channel_id
-AND name IN (sqlc.slice(':names'));
+AND user_id IN (sqlc.slice(':user_ids'));
 
 -- name: DeleteAllMatchStreamers :exec
 DELETE FROM streamers
@@ -26,7 +26,7 @@ WHERE channel_id = :channel_id;
 -- name: ListMatchStreamers :many
 SELECT
     channel_id,
-    name,
+    user_id,
     url
 FROM streamers
 WHERE channel_id = :channel_id;
